move sketches to separate file
